requirements.txt
# This is a normal Python requirements.txt file.

# Dependencies for image merging prediction
Pillow>=9.0.0